apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
    "PyJWT[crypto]>=2.9.0",
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "debugpy>=1.8.0",
    "snaptrade-python-sdk==11.0.132",
    "plaid-python>=36.1.0",
]

[project.scripts]
export-openapi = "scripts.export_openapi:main"
